+++ app/initializers/register-modal.js
@@ -1,1 +1,11 @@
-export { default, initialize } from 'ember-bootstrap-components/initializers/register-modal';
+import instanceInitializer from "../instance-initializers/register-modal";
+
+export default {
+  name: instanceInitializer.name,
+
+  initialize: function(registry, application) {
+    if (application.instanceInitializer) { return; }
+
+    instanceInitializer.initialize(application);
+  }
+};
